requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pydantic==2.6.0
pydantic-settings==2.1.0
slowapi==0.1.9
python-multipart==0.0.9
email-validator==2.1.0.post1
redis==5.0.1
alembic==1.13.1
greenlet
googlemaps==4.10.0
